requirements.txt
# requirements.txt

# Core dependencies
langchain==0.0.235
openai==0.27.0
tiktoken==0.3.3

# Vector database
faiss-cpu==1.7.3  # If using FAISS for indexing

# PDF processing
pypdf==3.8.1  # PDF loader for LangChain

# Environment variable handling
python-dotenv==1.0.0

# Testing
pytest==7.4.0

pdfplumber==0.9.0 # PDF text extraction